+++ src/index.ts
@@ -20,6 +20,7 @@ import { makeCosmosIbcPlugin } from './swap/defi/cosmosIbc'
 import { makeFantomSonicUpgradePlugin } from './swap/defi/fantomSonicUpgrade'
 import { makeLifiPlugin } from './swap/defi/lifi'
 import { makeRangoPlugin } from './swap/defi/rango'
+import { makeSwapsXyzSolanaPlugin } from './swap/defi/swapsxyzSolana'
 import { makeMayaProtocolPlugin } from './swap/defi/thorchain/mayaprotocol'
 import { makeSwapKitPlugin } from './swap/defi/thorchain/swapkit'
 import { makeSwapKitV3Plugin } from './swap/defi/thorchain/swapkitV3'
@@ -48,6 +49,7 @@ const plugins = {
   sideshift: makeSideshiftPlugin,
   spookySwap: makeSpookySwapPlugin,
   swapsxyz: makeSwapsXyzPlugin,
+  swapsxyzsolana: makeSwapsXyzSolanaPlugin,
   swapuz: makeSwapuzPlugin,
   mayaprotocol: makeMayaProtocolPlugin,
   thorchain: makeThorchainPlugin,